pertemuan-6

+++ rest-api/app/Http/Controllers/StudentController.php
@@ -12,7 +12,7 @@ class StudentController extends Controller
         # menggunakan model Student untuk select data
 		$students = Student::all();
 
-		if (!empty($students)) {
+		if ($students) {
 			$response = [
 				'message' => 'Menampilkan Data Semua Student',
 				'data' => $students,
@@ -22,8 +22,28 @@ class StudentController extends Controller
 			$response = [
 				'message' => 'Data tidak ada'
 			];
+			return response()->json($response, 400);
+		}
+	}
+
+	public function show($id){
+		$student = Student::find($id);
+
+		if($student){
+			$response = [
+				'message' => 'detail student',
+				'data' => $student,
+			];
+
 			return response()->json($response, 200);
 		}
+		else{
+			$response = [
+				'message' => 'Data student tidak ada'
+			];
+			return response()->json($response, 404);
+		}
+
 	}
 
 	public function store(Request $request) 
@@ -49,23 +69,40 @@ class StudentController extends Controller
 	public function update(Request $request, $id)
     {
 		$student = Student::find($id);
-		$student->update($request->all());
+		if($student){
+			
+			$student->update($request->all());
+			
+			$response = [
+				'message' => 'Data Student Berhasil diupdate',
+				'data' => $student,
+			];
+			return response()->json($response, 200);
+		}
+		else{
+			$response = [
+				'message' => 'Data student tidak ada'
+			];
+			return response()->json($response, 404);
+		}
+		
 
-		$response = [
-			'message' => 'Data Student Berhasil diupdate',
-			'data' => $student,
-		];
-		return response()->json($response, 201);
     }
 
 	public function destroy($id){
 		$student = Student::find($id);
-		$student->delete();
-		$response = [
-			'message' => 'Data Student Berhasil dihapus',
-			'data' => $student,
-		];
-
-		return response()->json($response, 201);
+		if($student){
+			$student->delete();
+			$response = [
+				'message' => 'Data student berhasil dihapus'
+			];
+			return response()->json($response, 200);
+		}
+		else{
+			$response = [
+				'message' => 'Data student tidak ada'
+			];
+			return response()->json($response, 404);
+		}
 	}
 }
